Create main.cpp
This program creates an online bookstore (requires input file in this form:
Story of Mars
Hanson, James
123A2345V  12.50 5)
Where a user is presented with a menu and can purchase a book, return a book, replace a book, view all items by a specific author, view all available books, view an alphabetized list of all books, view the details of a specific book, or exit and print the sales summary. It uses functions, 2 arrays, and structures.